Add additional check on rendering native preview panel
