+++ frontend/src/metabase/query_builder/components/view/ViewHeader/utils.ts
@@ -19,3 +19,24 @@ export const canExploreResults = (question: Question): boolean => {
     !question.isArchived()
   );
 };
+
+interface CanShowNativePreviewOpts {
+  question: Question;
+  queryBuilderMode: string;
+}
+
+export const canShowNativePreview = ({
+  question,
+  queryBuilderMode,
+}: CanShowNativePreviewOpts) => {
+  const { isNative } = Lib.queryDisplayInfo(question.query());
+  const isMetric = question.type() === "metric";
+
+  return (
+    !isNative &&
+    !isMetric &&
+    question.database()?.native_permissions === "write" &&
+    queryBuilderMode === "notebook" &&
+    !question.isArchived()
+  );
+};
